Refactor project structure and add Flask app.
Refactor the project by moving "game.py", "cards.py", and "poker.py" to the "core" directory. Modify import paths accordingly in other files. Introduce a basic Flask application with templates, static assets, and update "requirements.txt" to include new dependencies.
