+++ flask_app/static/style.css
@@ -1,0 +1,6 @@
+/* flask_app/static/style.css */
+body {
+    font-family: Arial, sans-serif;
+    text-align: center;
+    padding: 50px;
+}
