+++ flask_app/templates/index.html
@@ -1,0 +1,14 @@
+<!-- flask_app/templates/index.html -->
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>My Poker Face</title>
+    <link rel="stylesheet" href="{{ url_for('static', filename='style.css') }}">
+</head>
+<body>
+<h1>Welcome to Poker Night!</h1>
+<p>Let's play some poker.</p>
+</body>
+</html>
